requirements.txt
backoff~=1.10.0
google-api-python-client~=1.8.2
google-cloud-bigquery~=1.24.0
requests~=2.23.0